# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nova Elite FC Academy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  player: 'player',
  admin: 'admin',
} as const;

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  firstName: string;
  lastName: string;
  role: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  player: 'player',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  role: UserRole;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface PlayerInput {
  userId: number;
  dateOfBirth: string;
  position: string;
  nationality: string;
  phone: string;
  preferredFoot?: string;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  guardianName?: string | null;
  /** @nullable */
  guardianPhone?: string | null;
  /** @nullable */
  passportUrl?: string | null;
  /** @nullable */
  govIdUrl?: string | null;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  bio?: string | null;
}

export interface PlayerUpdate {
  position?: string;
  phone?: string;
  preferredFoot?: string;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  guardianName?: string | null;
  /** @nullable */
  guardianPhone?: string | null;
  /** @nullable */
  passportUrl?: string | null;
  /** @nullable */
  govIdUrl?: string | null;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  bio?: string | null;
  status?: string;
}

export interface Team {
  id: number;
  name: string;
  ageGroup: string;
  coach: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  playerCount?: number;
  trophies?: number;
  /** @nullable */
  winRate?: number | null;
}

export interface Player {
  id: number;
  userId: number;
  dateOfBirth: string;
  position: string;
  nationality: string;
  phone: string;
  preferredFoot?: string;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  guardianName?: string | null;
  /** @nullable */
  guardianPhone?: string | null;
  /** @nullable */
  passportUrl?: string | null;
  /** @nullable */
  govIdUrl?: string | null;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  bio?: string | null;
  status: string;
  user?: User;
  team?: Team;
  createdAt?: string;
}

export interface Program {
  id: number;
  title: string;
  description: string;
  ageRange: string;
  duration: string;
  level: string;
  /** @nullable */
  imageUrl?: string | null;
  features?: string[];
  /** @nullable */
  price?: number | null;
}

export interface TrialInput {
  title: string;
  date: string;
  location: string;
  ageGroup: string;
  spotsAvailable: number;
  /** @nullable */
  description?: string | null;
}

export interface Trial {
  id: number;
  title: string;
  date: string;
  location: string;
  ageGroup: string;
  spotsAvailable: number;
  spotsBooked: number;
  /** @nullable */
  description?: string | null;
  status?: string;
}

export interface TrialBookingInput {
  playerName: string;
  email: string;
  phone: string;
  dateOfBirth: string;
  position: string;
  /** @nullable */
  notes?: string | null;
}

export interface TrialBookingUpdate {
  status?: string;
}

export interface TrialBooking {
  id: number;
  trialId: number;
  playerName: string;
  email: string;
  phone: string;
  dateOfBirth: string;
  position: string;
  /** @nullable */
  notes?: string | null;
  status: string;
  createdAt?: string;
}

export interface NewsInput {
  title: string;
  content: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  excerpt?: string | null;
}

export interface NewsUpdate {
  title?: string;
  content?: string;
  category?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  excerpt?: string | null;
}

export interface NewsArticle {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  excerpt?: string | null;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  author?: string | null;
  createdAt: string;
}

export interface GalleryImageInput {
  imageUrl: string;
  title: string;
  category: string;
  /** @nullable */
  description?: string | null;
}

export interface GalleryImage {
  id: number;
  imageUrl: string;
  title: string;
  category: string;
  /** @nullable */
  description?: string | null;
  createdAt?: string;
}

export type SponsorTier = typeof SponsorTier[keyof typeof SponsorTier];


export const SponsorTier = {
  platinum: 'platinum',
  gold: 'gold',
  silver: 'silver',
  bronze: 'bronze',
} as const;

export interface Sponsor {
  id: number;
  name: string;
  logoUrl: string;
  /** @nullable */
  websiteUrl?: string | null;
  tier: SponsorTier;
}

export interface ScheduleInput {
  teamId: number;
  dayOfWeek: string;
  startTime: string;
  endTime: string;
  location: string;
  type: string;
  /** @nullable */
  notes?: string | null;
}

export interface TrainingSchedule {
  id: number;
  teamId: number;
  dayOfWeek: string;
  startTime: string;
  endTime: string;
  location: string;
  type: string;
  /** @nullable */
  notes?: string | null;
  team?: Team;
}

export interface ContactInput {
  name: string;
  email: string;
  subject: string;
  message: string;
  /** @nullable */
  phone?: string | null;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  subject: string;
  message: string;
  /** @nullable */
  phone?: string | null;
  status?: string;
  createdAt: string;
}

export interface ScoutingReportInput {
  playerName: string;
  age: number;
  position: string;
  nationality: string;
  scoutedAt: string;
  location: string;
  rating: number;
  notes: string;
  potential?: string;
  /** @nullable */
  playerId?: number | null;
}

export interface ScoutingReport {
  id: number;
  playerName: string;
  age: number;
  position: string;
  nationality: string;
  scoutedAt: string;
  location: string;
  rating: number;
  notes: string;
  potential?: string;
  /** @nullable */
  playerId?: number | null;
  createdAt?: string;
}

export type DashboardSummaryPlayersByTeamItem = {
  team: string;
  count: number;
};

export type DashboardSummaryBookingsByStatusItem = {
  status: string;
  count: number;
};

export interface DashboardSummary {
  totalPlayers: number;
  totalTrialBookings: number;
  pendingBookings: number;
  upcomingTrials: number;
  totalNews: number;
  recentMessages: number;
  playersByTeam?: DashboardSummaryPlayersByTeamItem[];
  bookingsByStatus?: DashboardSummaryBookingsByStatusItem[];
}

export interface PlayerDashboardSummary {
  player: Player;
  upcomingSchedules: TrainingSchedule[];
  /** @nullable */
  trialStatus: string | null;
  /** @nullable */
  teamName?: string | null;
}

export type ListPlayersParams = {
team?: string;
status?: string;
};

export type ListNewsParams = {
limit?: number;
};

export type ListGalleryParams = {
category?: string;
};

export type ListSchedulesParams = {
team?: string;
};