# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nova Elite FC Academy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trials:
    get:
      operationId: listTrials
      tags: [trials]
      summary: List upcoming trials
      responses:
        "200":
          description: Trials list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trial"
    post:
      operationId: createTrial
      tags: [trials]
      summary: Create a trial slot (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrialInput"
      responses:
        "201":
          description: Trial created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trial"
  /trials/{id}/book:
    post:
      operationId: bookTrial
      tags: [trials]
      summary: Book a trial slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrialBookingInput"
      responses:
        "201":
          description: Trial booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrialBooking"
...
components:
  schemas:
...
    TrialInput:
      type: object
      required: [title, date, location, ageGroup, spotsAvailable]
      properties:
        title:
          type: string
        date:
          type: string
        location:
          type: string
        ageGroup:
          type: string
        spotsAvailable:
          type: integer
        description:
          type: ["string", "null"]
    Trial:
      type: object
      required: [id, title, date, location, ageGroup, spotsAvailable, spotsBooked]
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
        location:
          type: string
        ageGroup:
          type: string
        spotsAvailable:
          type: integer
        spotsBooked:
          type: integer
        description:
          type: ["string", "null"]
        status:
          type: string
    TrialBookingInput:
      type: object
      required: [playerName, email, phone, dateOfBirth, position]
      properties:
        playerName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        position:
          type: string
        notes:
          type: ["string", "null"]
...
    TrialBooking:
      type: object
      required: [id, trialId, playerName, email, phone, dateOfBirth, position, status]
      properties:
        id:
          type: integer
        trialId:
          type: integer
        playerName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        position:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
...